refactor: 경로 이름 수정 entity->entities
