+++ server/src/app.module.ts
@@ -18,7 +18,7 @@ import { ChecklistAiModule } from './checklist-ai/checklist-ai.module';
 import { CommonModule } from './common/common.module';
 import { LoggingInterceptor } from './common/interceptor/log.interceptor';
 import { LoggerMiddleware } from './common/middlewares/logger.middleware';
-import { FeedModel } from './feeds/entity/feed.entity';
+import { FeedModel } from './feeds/entities/feed.entity';
 import { FeedsModule } from './feeds/feeds.module';
 import { FolderModel } from './folders/entities/folder.entity';
 import { FoldersModule } from './folders/folders.module';
