+++ server/src/common/entities/base.entity.ts
