+++ server/src/common/entities/checklist.entity.ts
