+++ server/src/feeds/feeds.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { FeedsService } from './feeds.service';
 import { FeedsController } from './feeds.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FeedModel } from './entity/feed.entity';
+import { FeedModel } from './entities/feed.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([FeedModel])],
