+++ server/src/feeds/feeds.service.spec.ts
@@ -2,7 +2,7 @@ import { BadRequestException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { FeedModel } from './entity/feed.entity';
+import { FeedModel } from './entities/feed.entity';
 import { FeedsService } from './feeds.service';
 
 type MockRepository<T = any> = Partial<Record<keyof Repository<T>, jest.Mock>>;
